fix: add a workaround for testcontainers/testcontainers-java#6441 (#152)
Adds a workaround for testcontainers/testcontainers-java#6441 issue.
The issue happens when Docker was unavailable during the first attempt to start a container. All future attempts will fail, since `testcontainers` caches the connection attempt result in a static variable, so the state will persist while the class is loaded.
The workaround is to reset the state through reflection each time before we start the container.